+++ packages/fxa-content-server/app/scripts/lib/experiments/grouping-rules/password-strength.js
@@ -5,14 +5,17 @@
 const BaseGroupingRule = require('./base');
 
 const GROUPS_FOR_PARTIAL_ROLLOUT = ['control', 'designF'];
-const ROLLOUT_RATE = 0.1;
+const ROLLOUT_RATES = {
+  'de': 0.2, // german is 10% control, 10% treatment
+  'en': 1.0, // english is rolled out to 50% control, 50% treatment
+};
 
 const EXPERIMENT_NAME = 'passwordStrength';
 
 class PasswordStrengthExperiment extends BaseGroupingRule {
   constructor () {
     super();
-    this.ROLLOUT_RATE = ROLLOUT_RATE;
+    this.ROLLOUT_RATES = ROLLOUT_RATES;
     this.name = EXPERIMENT_NAME;
   }
 
@@ -25,18 +28,21 @@ class PasswordStrengthExperiment extends BaseGroupingRule {
       return 'designF';
     }
 
-    if (! /^en/.test(subject.lang)) {
-      // Unless a tester, only available to english right now.
+    const rolloutLang = subject.lang.substr(0, 2);
+    const rolloutRate = this.ROLLOUT_RATES[rolloutLang];
+
+    if (! rolloutRate) {
       return false;
     }
 
-    if (this.bernoulliTrial(this.ROLLOUT_RATE, subject.uniqueUserId)) {
+    if (this.bernoulliTrial(rolloutRate, subject.uniqueUserId)) {
       return this.uniformChoice(GROUPS_FOR_PARTIAL_ROLLOUT, subject.uniqueUserId);
     }
 
     return false;
   }
 }
+
 
 PasswordStrengthExperiment.EXPERIMENT_NAME = EXPERIMENT_NAME;